backend/requirements.txt
Django>=4.2.14
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-celery-results
celery
redis
gunicorn
pandas
googlemaps
openai
requests
stripe
gunicorn==20.1.0
django-storages
python-dotenv==0.21.0
psycopg2-binary>=2.9
boto3
docker
psutil 
